+++ get_next_line_utils.c
@@ -1,10 +1,74 @@
 //funciones que puedes usar son (ft_strjoin) (ft_substr) (ft_strlen) (ft_strchr) (ft_strdup)
 //animo ;) ! i will stay here
+#include "get_next_line.h"
+
+size_t	ft_strlen(const char *s)
+{
+	size_t	i;
+
+	if(s == NULL)
+		return (0);
+	i = 0;
+	while (s[i])
+		i++;
+	return (i);
+}
+
+char	*ft_substr(char const *s, unsigned int start, size_t len)
+{
+	char	*new;
+	size_t	i;
+
+	i = 0;
+	if (s == NULL)
+		return (NULL);
+	if (start >= ft_strlen(s))
+		return (calloc(1,1));
+	if (len > ft_strlen(s) - start)
+		len = ft_strlen(s) - start;
+	new = calloc(sizeof(char), (len + 1));
+	if (new == NULL)
+		return (NULL);
+	while (i < len)
+	{
+		new[i] = s[start];
+		i++;
+		start++;
+	}
+	new[i] = '\0';
+	return (new);
+}
+char	*ft_strjoin(char const *s1, char const *s2)
+{
+	char	*new;
+	size_t	i;
+	size_t	len1;
+	size_t	len2;
+
+	i = 0;
+	if (s1 == NULL && s2 == NULL)
+		return (NULL);
+	len1 = ft_strlen(s1);
+	len2 = ft_strlen(s2);
+	new = calloc ((len1 + len2 + 1), sizeof(char));
+	if (new == NULL)
+		return (NULL);
+	while (i < (len1 + len2))
+	{
+		while (i < len1)
+			new[i++] = *s1++;
+		new[i++] = *s2++;
+	}
+	//new[i] = '\0'; 
+	return (new);
+}
 
 char	*ft_strchr(const char *s, int c)
 {
 	char	*ptr;
 
+	if (!s)
+		return (0);	
 	ptr = (char *)s;
 	while (*ptr != '\0')
 	{
@@ -16,3 +80,17 @@ char	*ft_strchr(const char *s, int c)
 		return (ptr);
 	return (0);
 }
+// char	*ft_strdup(const char *s1)
+// {
+// 	int		i;
+// 	char	*new;
+
+// 	i = 0;
+// 	new = calloc(ft_strlen(s1), sizeof(char) + 1);
+// 	if (new == NULL)
+// 		return (NULL);
+// 	while (*s1)
+// 		new[i++] = *s1++;
+// 	new[i] = '\0';
+// 	return (new);
+// }
